Unifica aplicaciones en un solo programa
